build: flip build script to bazel
